Move edit into the edit file and rename render.php into index.php
